requirements.txt
flask==3.0.0
requests==2.31.0
python-dotenv==1.0.0
apscheduler==3.10.4
TikTokLive
pytz>=2023.3; python_version<"3.9"

